feat: add font typeface-gfs-didot, move theme in separate file, delete local file didot.Достали его глюки
