+++ package.json
@@ -33,6 +33,7 @@
     "react": "^16.8.4",
     "react-dom": "^16.8.4",
     "react-helmet": "^5.2.0",
+    "typeface-gfs-didot": "0.0.72",
     "uuid": "^3.2.1"
   },
   "keywords": [
