+++ src/pages/style-guide.tsx
@@ -1,95 +1,22 @@
 import React from 'react';
-import { 
-  createMuiTheme,
-  useTheme,
+import {
   makeStyles, 
   createStyles, 
   Theme, 
-  withStyles,
-  responsiveFontSizes,
   ThemeProvider
 } from '@material-ui/core/styles';
-import { red } from '@material-ui/core/colors';
 import { 
   Typography, 
   Grid, 
   CssBaseline,
   Container,
-  Button,
   Divider,
-  AppBar
 } from '@material-ui/core';
 
 import Layout from '../components/Layout';
-import Navbar from '../components/Navbar';
-
-import Didot from '../../static/fonts/Didot.woff';
-
-const didot = {
-  fontFamily: 'Didot',
-  fontStyle: 'normal',
-  fontWeight: '400',
-  src: `local('Didot'),
-        url(${Didot}) format('woff')
-      `,
-};
-
-let theme = createMuiTheme({
-  palette: {
-    primary: {
-      light: '#E96F72',
-      main: '#E8676B',
-      dark: '#E96F72',
-      contrastText: '#FFFFFF',
-    },
-    secondary: {
-      main: '#FFFFFF',
-      dark: '#FFFFFF',
-      contrastText: '#3d4451',
-    },
-    text: {
-      primary: '#FFFFFF',
-      secondary: '#3d4451'
-    },
-    background: {
-      default: '#FFFFFF',
-    },
-  },  typography: {
-    fontFamily: 'Didot',
-    fontSize: 16,
-    h1: {
-      fontSize: '2.25rem',
-      lineHeight: '1.2',
-      fontWeight: 'bold',
-      color: '#3d4451',
-      margin: '10px 0',
-    },
-    h2: {
-      fontSize: '1.875rem',
-      lineHeight: '1.2',
-      fontWeight: 'bold',
-      color: '#3d4451',
-      margin: '10px 0',
-    },
-    body1: {
-      fontSize: '1.25rem',
-      color: '#000000',
-    },
-    body2: {
-      fontSize: '1rem',
-      color: 'rgb(61, 68, 81);',
-    },
-  },
-  spacing: factor => `${0.25 * factor}rem`, // (Bootstrap strategy)
-  overrides: {
-    MuiCssBaseline: {
-      '@global': {
-        '@font-face': [didot],
-      },
-    },
-  },
-});
-theme = responsiveFontSizes(theme);
+import ResponsiveDrawer from '../components/Navbar/Navbar';
+import { ColoredButton, OutlinedButton } from '../components/Buttons'
+import theme from '../../static/theme';
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
@@ -136,26 +63,6 @@ const Typos = () => {
   )
 }
 
-const ColorButton = withStyles((theme: Theme) => ({
-  root: {
-    '&:hover': {
-      boxShadow: '0 5px 11px 0 rgba(0, 0, 0, 0.18), 0 4px 15px 0 rgba(0, 0, 0, 0.15)',
-    },
-  },
-}))(Button);
-
-const OutlinedButton = withStyles((theme: Theme) => ({
-  root: {
-    color: '#3d4451',
-    border: '1px solid #cbcdcf',
-    '&:hover': {
-      color: '#3d4451',
-      border: '1px solid #cbcdcf',
-      boxShadow: '0 5px 11px 0 rgba(0, 0, 0, 0.18), 0 4px 15px 0 rgba(0, 0, 0, 0.15)',
-    },
-  },
-}))(Button);
-
 const Buttons = () => {
   const classes = useStyles();
   return (
@@ -163,11 +70,9 @@ const Buttons = () => {
       <Typography variant="h2">Buttons</Typography>      
       <Divider light/>
       <div className={classes.root}>
-        <Button variant="contained" color="primary">Primary Default</Button>
-        <Button variant="contained" color="secondary">secondary Default</Button>
-        <ColorButton variant="contained" color="primary" size="medium">
+        <ColoredButton variant="contained" color="primary" size="medium">
           Primary
-        </ColorButton>
+        </ColoredButton>
         <OutlinedButton variant="outlined" size="medium" color="secondary">
           Outlined
         </OutlinedButton>
@@ -181,24 +86,24 @@ const NavBars = () => {
     <Container maxWidth="lg">
       <Typography variant="h2">Navbar</Typography>
       <Divider light/>
-      <Navbar />
+      <ResponsiveDrawer />
     </Container>
   )
 }
 const StyleGuide = () => (
-  <Layout>
-    <CssBaseline />
-    <ThemeProvider theme={theme}>
-      <Container maxWidth="lg">
-        <h1>StyleGuide</h1>
-        <Grid container spacing={3}>
-          <Buttons />          
-          <Typos />
-          <NavBars />
-        </Grid>
-      </Container>
-    </ThemeProvider>
-  </Layout>
+  <ThemeProvider theme={theme}>
+    <Layout>
+      <CssBaseline />
+        <Container maxWidth="lg">
+          <h1>StyleGuide</h1>
+          <Grid container spacing={3}>
+            <Buttons />          
+            <Typos />
+            <NavBars />
+          </Grid>
+        </Container>
+    </Layout>
+  </ThemeProvider>
 );
 
 export default StyleGuide;
